front fixes

+++ blog-backend/BlogAPI/Controllers/PostsController.cs
@@ -1,92 +1,86 @@
 ﻿using BlogAPI.DTOs.Posts;
-using BlogAPI.Models;
+using BlogAPI.Helpers;
 using BlogAPI.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using MongoDB.Bson;
 
 namespace BlogAPI.Controllers
 {
     [ApiController]
-    [Authorize(Roles = "author")]
     [Route("api/[controller]")]
-
     public class PostsController : BaseController
     {
         private readonly PostsService _postsService;
+        private readonly UsersService _usersService;
+        private readonly AuthorsService _authorsService;
 
-        public PostsController(PostsService postsService, IWebHostEnvironment env) : base(env)
+
+        public PostsController(IWebHostEnvironment env, PostsService postsService, UsersService usersService, AuthorsService authorsService) : base(env)
         {
             _postsService = postsService;
+            _usersService = usersService;
+            _authorsService = authorsService;
         }
 
-        [HttpGet("GetApprovedPosts")]
         [AllowAnonymous]
-        public async Task<ActionResult<List<Posts>>> GetApprovedPosts()
+        [HttpGet("GetImage")]
+        public async Task<ActionResult> GetImage(string id)
+        {
+            try
+            {
+                var post = await _postsService.GetPostByMainPhotoIdAsync(id);
+
+                if (post == null)
+                {
+                    throw new ArgumentException("Post using that image not found");
+                }
+
+                var result = await _postsService.GetPostImageAsync(id);
+
+                if (post.Status == Status.Aproved)
+                {
+                    return File(result, "image/jpeg");
+                }
+
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+
+                if (post.AuthorId != authorId)
+                    throw new UnauthorizedAccessException();
+
+                return File(result, "image/jpeg");
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+        [AllowAnonymous]
+        [HttpGet("GetApprovedPosts")]
+        public async Task<ActionResult<List<PostsDTO>>> GetApprovedPosts()
         {
             try
             {
                 var result = await _postsService.GetPostsByStatusAsync(Status.Aproved);
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return HandleError(ex);
-            }
-        }
 
-        [HttpGet("GetDraftPosts")]
-        public async Task<ActionResult<List<Posts>>> GetDraftPosts()
-        {
-            try
-            {
-                var userId = User.Identity.Name;
+                var res = result.Select(post => new PostsDTO
+                {
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
 
-                var result = await _postsService.GetPostsByStatusAndUserIdAsync(Status.Draft, userId);
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return HandleError(ex);
-            }
-        }
 
-        [HttpGet("GetRejectedPosts")]
-        public async Task<ActionResult<List<Posts>>> GetRejectedPosts()
-        {
-            try
-            {
-                var result = await _postsService.GetPostsByStatusAndUserIdAsync(Status.Rejected, "");
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return HandleError(ex);
-            }
-        }
-
-        [HttpGet("GetPostWaitingForApproval")]
-        [Authorize(Roles = "admin")]
-        public async Task<ActionResult<List<Posts>>> GetPostWaitingForApproval()
-        {
-            try
-            {
-                var result = await _postsService.GetPostsByStatusAsync(Status.ToConfirm);
-                return Ok(result);
-            }
-            catch (Exception ex)
-            {
-                return HandleError(ex);
-            }
-        }
-
-        [HttpGet("GetApprovedPostsByAuthorId")]
-        public async Task<ActionResult<Posts>> GetApprovedPostsByAuthorId(string id)
-        {
-            try
-            {
-
-                var res = await _postsService.GetApprovedPostsByAuthorIdAsync(id);
                 return Ok(res);
             }
             catch (Exception ex)
@@ -95,28 +89,221 @@ namespace BlogAPI.Controllers
             }
         }
 
-        [HttpPost("CreatePost")]
-        public async Task<ActionResult> CreatePost(PostsDTO newPost)
+        [Authorize("write:posts")]
+        [HttpGet("GetDraftPosts")]
+        public async Task<ActionResult<List<PostsDTO>>> GetDraftPosts()
         {
             try
             {
-                if (User?.Identity?.Name != newPost.AuthorId)
-                    throw new UnauthorizedAccessException();
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
 
-                Posts post = new()
+                var result = await _postsService.GetPostsByStatusAndAuthorIdAsync(Status.Draft, authorId);
+
+                var res = result.Select(post => new PostsDTO
                 {
-                    AuthorId = ObjectId.Parse(newPost.AuthorId),
-                    Category = newPost.Category,
-                    Content = newPost.Content,
-                    CreatedAt = DateTime.UtcNow,
-                    Id = MongoDB.Bson.ObjectId.GenerateNewId(),
-                    MainParentId = null,
-                    Status = Status.Draft,
-                    Tags = newPost.Tags,
-                    Title = newPost.Title
-                };
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
 
-                await _postsService.CreatePostAsync(post);
+                return Ok(res);
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+        [Authorize("write:posts")]
+        [HttpGet("GetRejectedPosts")]
+        public async Task<ActionResult<List<PostsDTO>>> GetRejectedPosts()
+        {
+            try
+            {
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+                var result = await _postsService.GetPostsByStatusAndAuthorIdAsync(Status.Rejected, authorId);
+
+                var res = result.Select(post => new PostsDTO
+                {
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
+
+                return Ok(res);
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+
+        [Authorize("write:posts")]
+        [HttpGet("GetPostWaitingForApproval")]
+        public async Task<ActionResult<List<PostsDTO>>> GetPostWaitingForApproval()
+        {
+            try
+            {
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+                var result = await _postsService.GetPostsByStatusAndAuthorIdAsync(Status.ToConfirm, authorId);
+
+                var res = result.Select(post => new PostsDTO
+                {
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
+
+                return Ok(res);
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+        [AllowAnonymous]
+        [HttpGet("GetApprovedPostsByAuthorId")]
+        public async Task<ActionResult<PostsDTO>> GetApprovedPostsByAuthorId(string authorId)
+        {
+            try
+            {
+                var result = await _postsService.GetApprovedPostsByAuthorIdAsync(authorId);
+
+                var res = result.Select(post => new PostsDTO
+                {
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
+
+                return Ok(res);
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+
+        [AllowAnonymous]
+        [HttpGet("GetApprovedPostsByCategory")]
+        public async Task<ActionResult<List<PostsDTO>>> GetApprovedPostsByCategory(string category)
+        {
+            try
+            {
+                var result = await _postsService.GetApprovedPostsByCategoryAsync(category);
+
+                var res = result.Select(post => new PostsDTO
+                {
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
+
+                return Ok(res);
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+        [AllowAnonymous]
+        [HttpGet("GetApprovedPostsByTag")]
+        public async Task<ActionResult<List<PostsDTO>>> GetApprovedPostsByTag(string tag)
+        {
+            try
+            {
+                var result = await _postsService.GetApprovedPostsByTagAsync(tag);
+
+                var res = result.Select(post => new PostsDTO
+                {
+                    Id = post.Id.ToString(),
+                    Title = post.Title,
+                    Category = post.Category,
+                    Content = post.Content,
+                    Tags = post.Tags,
+                    MainPhotoId = post.MainPhotoId.ToString(),
+                    AuthorId = post.AuthorId.ToString(),
+                    CreatedAt = post.CreatedAt,
+                    Status = post.Status,
+                    Dislikes = post.Dislikes,
+                    Likes = post.Likes,
+                    Views = post.Views,
+                    MainParentId = post.MainParentId?.ToString()
+                }).ToList();
+
+                return Ok(res);
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+        [Authorize("write:posts")]
+        [HttpPost("CreateDraftPost")]
+        public async Task<ActionResult> CreateDraftPost([FromForm] CreateDraftPostDTO newDraftPost)
+        {
+
+            try
+            {
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+
+                await _postsService.CreateDraftPostAsync(newDraftPost, Status.Draft, authorId);
+
                 return Ok("Post created");
             }
             catch (Exception ex)
@@ -125,19 +312,42 @@ namespace BlogAPI.Controllers
             }
         }
 
+
+        [Authorize("write:posts")]
+        [HttpPost("CreatePost")]
+        public async Task<ActionResult> CreatePost([FromForm] CreatePostDTO newPost)
+        {
+
+            try
+            {
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+
+                await _postsService.CreatePostAsync(newPost, Status.ToConfirm, authorId);
+
+                return Ok("Post created");
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+
+        [Authorize("write:posts")]
         [HttpPut("UpdateDraftPost")]
-        [Authorize(Policy = "author")]
-        public async Task<ActionResult> UpdateDraftPost(PostsDTO postDto)
+        public async Task<ActionResult> UpdateDraftPost(UpdatePostDTO postDto)
         {
             try
             {
-                if (User?.Identity?.Name != postDto.AuthorId)
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+                var post = await _postsService.GetPostByIdAsync(postDto.Id);
+
+                if (post?.AuthorId != authorId)
                     throw new UnauthorizedAccessException();
 
-                if (postDto?.Id != null)
-                {
-                    await _postsService.UpdateDraftPostAsync(postDto);
-                }
+
+                await _postsService.UpdateDraftPostAsync(postDto);
+
                 return Ok("Post updated");
             }
             catch (Exception ex)
@@ -148,14 +358,38 @@ namespace BlogAPI.Controllers
 
 
 
+        [Authorize("write:posts")]
+        [HttpPut("UpdateAcceptedPost")]
+        public async Task<ActionResult> UpdateAcceptedPost(UpdatePostDTO postDto)
+        {
+            try
+            {
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+                var post = await _postsService.GetPostByIdAsync(postDto.Id);
+
+                if (post?.AuthorId != authorId)
+                    throw new UnauthorizedAccessException();
+
+
+                await _postsService.UpdateAcceptedPostAsync(postDto);
+
+                return Ok("Post updated");
+            }
+            catch (Exception ex)
+            {
+                return HandleError(ex);
+            }
+        }
+
+
         [HttpPut("AcceptPost")]
-        [Authorize(Policy = "admin")]
+        [Authorize("admin")]
         public async Task<ActionResult> AcceptPost(string id)
         {
             try
             {
-                await _postsService.RejectPostAsync(id);
-                return Ok("Post updated");
+                await _postsService.AcceptPostAsync(id);
+                return Ok("Post accepted");
             }
             catch (Exception ex)
             {
@@ -164,7 +398,7 @@ namespace BlogAPI.Controllers
         }
 
         [HttpPut("RejectPost")]
-        [Authorize(Policy = "admin")]
+        [Authorize("admin")]
         public async Task<ActionResult> RejectPost(string id)
         {
             try
@@ -178,12 +412,18 @@ namespace BlogAPI.Controllers
             }
         }
 
+        [Authorize("delete:posts")]
         [HttpDelete("DeletePost")]
-        [Authorize(Policy = "author")]
         public async Task<ActionResult> DeletePost(string id)
         {
             try
             {
+                var authorId = IdHelper.GetAuthorId(User, _usersService, _authorsService);
+                var post = await _postsService.GetPostByIdAsync(id);
+
+                if (post?.AuthorId != authorId)
+                    throw new UnauthorizedAccessException();
+
                 await _postsService.DeletePostAsync(id);
                 return Ok("Post deleted");
             }
